Noise

Noise

Example Julia package repo.

Noise.additive_white_gaussian — Function.

additive_white_gaussian(X; clip=false[, σ=1.0, μ=0.0])

Returns the array X with Gaussian noise (standard deviation σ and mean μ) added. σ and μ are optional arguments representing standard deviation and mean of Gaussian. If keyword argument clip is provided the values are clipped to be in [0, 1].

source

Noise.additive_white_gaussian_chn — Function.

additive_white_gaussian_chn(X[, σ=1.0, μ=0.0])

Returns the RGB image X with Gaussian noise (standard deviation σ and mean μ) added pixelwise. However, every channel of one pixel receives the same amount of noise. The noise therefore acts roughly as intensity - but not color - changing noise. σ and μ are optional arguments representing standard deviation and mean of Gaussian.

source

Noise.poisson — Function.

poisson(X; scaling=1, clip=false)

Returns the array X affected by Poisson noise. At every position the Poisson noise affects the intensity individually and the values at the positions represent the expected value of the Poisson Distribution.

scaling is a optional argument to scale the highest value of the array to scaling. This can be used if the maximum intensity corresponds to a number of photons for example.

clip is a keyword argument. If given, it clips the values to [0, 1]

Please keep in mind that a Poisson Distribution returns integers therefore a image with intensity [0, 1] and scaling=1.0 will only return integers (0, 1, 2, ...).

source

Noise.salt_pepper — Function.

salt_pepper(X; salt_prob=0.5, salt=1.0, pepper=0.0[, prob])

Returns array X affected by salt and pepper noise. prob is a optional argument for the probability that a pixel will be affected by the noise. salt_prob is a keyword argument representing the probability for salt noise. The probability for pepper noise is therefore 1-salt_prob. salt is a keyword argument for specifying the value of salt noise. pepper is a keyword argument for specifying the value of pepper noise.

source

Noise.salt_pepper_chn — Function.

salt_pepper_chn(X; salt_prob=0.5, salt=1.0, pepper=0.0[, prob])

Returns a RGB Image X affected by salt and pepper noise. When a salt or pepper occurs, it is applied to all channels of the RGB making a real salt and pepper on the whole image. prob is a optional argument for the probability that a pixel will be affected by the noise. salt_prob is a keyword argument representing the probability for salt noise. The probability for pepper noise is therefore 1-salt_prob. salt is a keyword argument for specifying the value of salt noise. pepper is a keyword argument for specifying the value of pepper noise.

source