Noise

Noise

Example Julia package repo.

Noise.additive_white_gaussian — Function.

additive_white_gaussian(X; clip=false[, σ=1.0, μ=0.0])

Returns the array X with Gaussian noise (standard deviation σ and mean μ) added. σ and μ are optional arguments representing standard deviation and mean of Gaussian. If keyword argument clip is provided the values are clipped to be in [0, 1]. If X is a RGB{Normed} or Gray{Normed} image, then the values will be automatically clipped and the keyword clip is meaningless.

source

Noise.additive_white_gaussian_chn — Function.

additive_white_gaussian_chn(X[, σ=1.0, μ=0.0])

Returns the RGB image X with Gaussian noise (standard deviation σ and mean μ) added pixelwise. However, every channel of one pixel receives the same amount of noise. The noise therefore acts roughly as intensity - but not color - changing noise. σ and μ are optional arguments representing standard deviation and mean of Gaussian.

source

Noise.poisson — Function.

poisson(X; scaling=1, clip=false)

Returns the array X affected by Poisson noise. At every position the Poisson noise affects the intensity individually and the values at the positions represent the expected value of the Poisson Distribution. Since Poisson Noise due to discrete events you should provide the optional argument scaling. This scaling connects the highest value of the array with the discrete number of events. The highest value will be then scaled and the poisson noise is applied Afterwards we scale the whole array back so that the initial intensity is preserved but with applied Poisson noise. clip is a keyword argument. If given, it clips the values to [0, 1]

source

Noise.salt_pepper — Function.

salt_pepper(X; salt_prob=0.5, salt=1.0, pepper=0.0[, prob])

Returns array X affected by salt and pepper noise. X can be an array or an RGB or Gray image prob is a optional argument for the probability that a pixel will be affected by the noise. salt_prob is a keyword argument representing the probability for salt noise. The probability for pepper noise is therefore 1-salt_prob. salt is a keyword argument for specifying the value of salt noise. pepper is a keyword argument for specifying the value of pepper noise.

source

Noise.salt_pepper_chn — Function.

salt_pepper_chn(X; salt_prob=0.5, salt=1.0, pepper=0.0[, prob])

Returns a RGB Image X affected by salt and pepper noise. When a salt or pepper occurs, it is applied to all channels of the RGB making a real salt and pepper on the whole image. prob is a optional argument for the probability that a pixel will be affected by the noise. salt_prob is a keyword argument representing the probability for salt noise. The probability for pepper noise is therefore 1-salt_prob. salt is a keyword argument for specifying the value of salt noise. pepper is a keyword argument for specifying the value of pepper noise.

source